pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aws-docs-to-epub"
version = "1.0.0"
description = "Convert AWS Developer Guide documentation to EPUB format"
authors = [{ name = "Brian Heise", email = "brian@example.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
keywords = ["aws", "documentation", "epub", "converter", "ebook"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
]

dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "ebooklib>=0.18",
    "Pillow>=10.0.0",
    "cairosvg>=2.7.0",
    "lxml>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pylint>=2.17.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/aws-docs-to-epub"
Repository = "https://github.com/yourusername/aws-docs-to-epub"
Issues = "https://github.com/yourusername/aws-docs-to-epub/issues"

[project.scripts]
aws-docs-to-epub = "aws_docs_to_epub.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=aws_docs_to_epub --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
